[refactor] 정보를 새로고침하는 기능 구현 및 ViewController 파일 분리
- collectionview의 refreshControl 할당
- ViewController가독성 개선을 위한 파일 분리
