+++ WeatherForecast/WeatherForecast/Controller/ViewController.swift
@@ -9,15 +9,16 @@ import CoreLocation
 
 class ViewController: UIViewController {
     private let locationManager = CLLocationManager()
-    private var currentWeather: WeatherData?
-    private var forecastWeather: [WeatherData]?
-    private var userAddress: String?
-    private var collectionView: UICollectionView!
+    var currentWeather: WeatherData?
+    var forecastWeather: [WeatherData]?
+    var userAddress: String?
+    private(set) var collectionView: UICollectionView!
     
     override func viewDidLoad() {
         super.viewDidLoad()
         configureLocationManager()
         setUpCollectionView()
+        configureRefreshControl()
     }
 
     private func configureLocationManager() {
@@ -53,98 +54,16 @@ class ViewController: UIViewController {
         collectionView.register(WeatherHeaderView.self, forSupplementaryViewOfKind: UICollectionView.elementKindSectionHeader, withReuseIdentifier: WeatherHeaderView.id)
         collectionView.register(ForecastWeatherCell.self, forCellWithReuseIdentifier: ForecastWeatherCell.id)
     }
-}
-
-extension ViewController: UICollectionViewDataSource {
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        guard let count = forecastWeather?.count else {
-            return 40
-        }
-
-        return count
+    
+    private func configureRefreshControl() {
+        let control = UIRefreshControl()
+        control.addTarget(self, action: #selector(refreshCollectionView), for: .valueChanged)
+        collectionView.refreshControl = control
     }
     
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        let cell = collectionView.dequeueReusableCell(withReuseIdentifier: ForecastWeatherCell.id, for: indexPath) as! ForecastWeatherCell
-        let data = forecastWeather?[indexPath.row]
-        cell.updateWeather(data)
-
-        return cell
-    }
-    
-    func collectionView(_ collectionView: UICollectionView, viewForSupplementaryElementOfKind kind: String, at indexPath: IndexPath) -> UICollectionReusableView {
-        let header = collectionView.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: WeatherHeaderView.id, for: indexPath) as! WeatherHeaderView
-        header.updateWeather(currentWeather, in: userAddress)
-
-        return header
+    @objc private func refreshCollectionView() {
+        locationManager.requestLocation()
+        collectionView.refreshControl?.endRefreshing()
     }
 }
 
-extension ViewController: CLLocationManagerDelegate {
-    func locationManagerDidChangeAuthorization(_ manager: CLLocationManager) {
-        switch manager.authorizationStatus {
-        case .authorizedWhenInUse, .authorizedAlways:
-            manager.requestLocation()
-        default:
-            manager.requestWhenInUseAuthorization()
-        }
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didUpdateLocations locations: [CLLocation]) {
-        guard let location = locations.first else {
-            return
-        }
-
-        let coordinate = CurrentCoordinate(of: location)
-
-        Task {
-            updateAddress(to: location) {
-                $0.formatAddress()
-            }
-
-            try await updateCurrentWeather(for: coordinate)
-            try await updateForecastWeather(for: coordinate)
-            collectionView.reloadData()
-        }
-    }
-    
-    func locationManager(_ manager: CLLocationManager, didFailWithError error: Error) {
-        print(WeatherNetworkError.requestFailed("위치").description)
-    }
-    
-    private func updateCurrentWeather(for location: CurrentCoordinate) async throws {
-        let current = try await WeatherParser.parseData(at: location, type: CurrentWeatherComponents.self)
-        currentWeather = WeatherData(current: current)
-        try await currentWeather?.convertToImage {
-            self.currentWeather?.iconImage = $0
-        }
-    }
-    
-    private func updateForecastWeather(for location: CurrentCoordinate) async throws {
-        let forecast = try await WeatherParser.parseData(at: location, type: ForecastWeatherComponents.self)
-        forecastWeather = forecast.list.map { WeatherData(forecast: $0) }
-
-        guard let forecastWeather else {
-            return
-        }
-
-        for (index, weatherData) in forecastWeather.enumerated() {
-            var image: UIImage?
-            try await weatherData.convertToImage {
-                image = $0
-            }
-
-            self.forecastWeather?[index].iconImage = image
-        }
-    }
-
-    private func updateAddress(to location: CLLocation, _ completion: @escaping (CLPlacemark) -> String?) {
-        CLGeocoder().reverseGeocodeLocation(location) { places, _ in
-            guard let place = places?.first else {
-                return
-            }
-
-            self.userAddress = completion(place)
-        }
-    }
-}
